% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_zinmf.R
\name{cosineDist}
\alias{cosineDist}
\title{Helper function to calculate cosine distance between columns of two matrices.}
\usage{
cosineDist(x, y)
}
\arguments{
\item{x}{A matrix of dimensions \eqn{P \times N_1}.}

\item{y}{A matrix of dimensions \eqn{P \times N_2}.}
}
\value{
A matrix containing the pairwise cosine distances between x and y.
}
\description{
Helper function to calculate cosine distance between columns of two matrices.
}
