% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_zinmf.R
\name{initialize_zinmf}
\alias{initialize_zinmf}
\title{Generate hyperparameters for gibbs sampling.}
\usage{
initialize_zinmf(vec = list())
}
\arguments{
\item{vec}{A named list containing custom hyperparameter values. Any missing components will be filled in with the default value. The components are:
\describe{
\item{alpha_w}{Shape parameter for p(w_{pk}). Default is 1.}
\item{beta_w}{Rate parameter for p(w_{pk}). Default is 5.}
\item{c_h}{Shape parameter for non-null scores. Default is 10.}
\item{epsilon}{Mean for null scores. Default is 0.25.}
\item{a_mu}{Shape parameter for p(mu). Default is 2.}
\item{b_mu}{Scale parameter for p(mu). Default is 20.}
\item{a_a}{Alpha parameter for p(pi_ps). Default is 1.}
\item{b_a}{Beta parameter for p(pi_pis). Default is 1.}
\item{beta_0}{Prior mean for intercept of probit stick breaking parameter. Default is 1.5.}
\item{tau_0}{Prior precision for probit stick breaking parameters. DEfault is 2.}
}}
}
\value{
A list containing hyperparameters to use with the gibbs sampler.
}
\description{
Generate hyperparameters for gibbs sampling.
}
\examples{

#Use Default hyperparameters
hyperparams <- initialize_zinmf()

# Change alpha_w parameter:
hyperparams_2 <- initialize_zinmf(list('alpha_w'=2))

}
